+++ classes/Sorting.php
@@ -40,6 +40,32 @@ class Sorting {
 			return 0;
 		return $a->getDate() < $b->getDate() ? 1 : -1;
 	}
+
+	/**
+	 * Sorts two DownloadFile by name, ascending oder.
+	 * 
+	 * @param DownloadFile $a first item
+	 * @param DownloadFile $b second item
+	 * @return int
+	 */
+	public static function downloadFileByNameAsc($a, $b) {
+		$a = strtolower($a->getName());
+		$b = strtolower($b->getName());
+		return strcmp($a, $b);
+	}
+
+	/**
+	 * Sorts two DownloadFile by name, descending oder.
+	 * 
+	 * @param DownloadFile $a first item
+	 * @param DownloadFile $b second item
+	 * @return int
+	 */
+	public static function downloadFileByNameDesc($a, $b) {
+		$a = strtolower($a->getName());
+		$b = strtolower($b->getName());
+		return -1 * strcmp($a, $b);
+	}
 }
 
 ?>